test(federation): split receiver-realm coverage by concern
