+++ tests/unit/test_kamiwaza_sdk_services_federations.py
@@ -766,274 +766,3 @@ def test_disconnect_force_passes_param() -> None:
     FederationsAPI(client)["LYRA"].disconnect(force=True)
     post = [kw for m, p, kw in client.calls if m == "POST"][0]
     assert post.get("params") == {"force": "true"}
-
-
-# ---------------------------------------------------------------------------
-# ENG-8213 Alt D (receiver_realm) — realm_scope on pair() + guest sub-resource
-# ---------------------------------------------------------------------------
-
-
-def test_pair_forwards_realm_scope_when_supplied() -> None:
-    """Supplying ``realm_scope`` creates the federation in the receiver-owned
-    receiver_realm mode (Alt D / design section 15): the receiver provisions a
-    dedicated ``federation-<id>`` realm and mints its own guest credentials. The SDK
-    forwards ``realm_scope`` on the create body."""
-    from kamiwaza_sdk.services.federations import FederationsAPI
-
-    client = _MockClient()
-    _stage_pair_responses(client)
-
-    api = FederationsAPI(client)
-    api.pair(
-        name="ORION",
-        role="receiver",
-        remote_url="https://orion.example.com",
-        realm_scope="per_federation",
-    )
-
-    _, body = _create_call(client)
-    assert body["realm_scope"] == "per_federation"
-
-
-def test_pair_omits_realm_scope_when_not_supplied() -> None:
-    """Without ``realm_scope`` the create body carries no realm_scope key (the
-    federation stays peer_kc / shared_idp per the other inputs)."""
-    from kamiwaza_sdk.services.federations import FederationsAPI
-
-    client = _MockClient()
-    _stage_pair_responses(client)
-
-    api = FederationsAPI(client)
-    api.pair(name="ORION", role="initiator", remote_url="https://orion.example.com")
-
-    _, body = _create_call(client)
-    assert "realm_scope" not in body
-
-
-def _stage_guest_federation(client: _MockClient, *, fid: str) -> None:
-    """Stage the GET list used to resolve a federation name -> id for guest ops."""
-    client.expect(
-        "GET",
-        "/cluster/federations",
-        [{"id": fid, "status": "PAIRED", "remote_cluster_name": "ORION"}],
-    )
-
-
-def test_guests_property_returns_guests_api() -> None:
-    from kamiwaza_sdk.services.federations import (
-        FederationGuestsAPI,
-        FederationsAPI,
-    )
-
-    client = _MockClient()
-    guests = FederationsAPI(client)["ORION"].guests
-    assert isinstance(guests, FederationGuestsAPI)
-
-
-def test_guests_enroll_posts_and_returns_credential() -> None:
-    """``guests.enroll(external_id)`` POSTs to the receiver's /guests endpoint
-    and returns the minted offline credential (returned ONCE, design 15.2)."""
-    from kamiwaza_sdk.services.federations import FederationsAPI
-
-    client = _MockClient()
-    fid = "11111111-1111-1111-1111-111111111111"
-    _stage_guest_federation(client, fid=fid)
-    client.expect(
-        "POST",
-        f"/cluster/federations/{fid}/guests",
-        {
-            "external_id": "carol@src-uuid",
-            "realm": f"federation-{fid}",
-            "offline_token": "OFFLINE-CRED-XYZ",
-        },
-    )
-
-    guest = FederationsAPI(client)["ORION"].guests.enroll("carol@src-uuid")
-
-    assert guest.external_id == "carol@src-uuid"
-    assert guest.realm == f"federation-{fid}"
-    assert guest.offline_token == "OFFLINE-CRED-XYZ"
-    post = [(p, kw) for m, p, kw in client.calls if m == "POST"][0]
-    assert post[0] == f"/cluster/federations/{fid}/guests"
-    assert post[1].get("json") == {"external_id": "carol@src-uuid"}
-
-
-def test_guests_enroll_forwards_initial_tuples() -> None:
-    """initial_tuples seed the guest's ReBAC grants at enrollment; the SDK
-    forwards them on the body only when supplied."""
-    from kamiwaza_sdk.services.federations import FederationsAPI
-
-    client = _MockClient()
-    fid = "22222222-2222-2222-2222-222222222222"
-    _stage_guest_federation(client, fid=fid)
-    client.expect(
-        "POST",
-        f"/cluster/federations/{fid}/guests",
-        {"external_id": "dave@src", "realm": f"federation-{fid}", "offline_token": "T"},
-    )
-
-    tuples = [{"subject": "user:dave", "relation": "reader", "object": "dataset:x"}]
-    FederationsAPI(client)["ORION"].guests.enroll("dave@src", initial_tuples=tuples)
-
-    post = [kw for m, p, kw in client.calls if m == "POST"][0]
-    assert post["json"]["initial_tuples"] == tuples
-
-
-def test_guests_enroll_forwards_identity_proof() -> None:
-    """The out-of-band identity_proof (design §7.4) is forwarded on the body only
-    when supplied — the manual default needs nothing, mtls needs a cert."""
-    from kamiwaza_sdk.services.federations import FederationsAPI
-
-    client = _MockClient()
-    fid = "44444444-4444-4444-4444-444444444444"
-    _stage_guest_federation(client, fid=fid)
-    client.expect(
-        "POST",
-        f"/cluster/federations/{fid}/guests",
-        {"external_id": "eve@src", "realm": f"federation-{fid}", "offline_token": "T"},
-    )
-
-    proof = {
-        "client_cert_pem": "-----BEGIN CERTIFICATE-----\nX\n-----END CERTIFICATE-----"
-    }
-    FederationsAPI(client)["ORION"].guests.enroll("eve@src", identity_proof=proof)
-
-    post = [kw for m, p, kw in client.calls if m == "POST"][0]
-    assert post["json"]["identity_proof"] == proof
-
-
-def test_guests_enroll_omits_identity_proof_when_not_supplied() -> None:
-    from kamiwaza_sdk.services.federations import FederationsAPI
-
-    client = _MockClient()
-    fid = "55555555-5555-5555-5555-555555555555"
-    _stage_guest_federation(client, fid=fid)
-    client.expect(
-        "POST",
-        f"/cluster/federations/{fid}/guests",
-        {"external_id": "frank@src", "realm": f"federation-{fid}", "offline_token": "T"},
-    )
-
-    FederationsAPI(client)["ORION"].guests.enroll("frank@src")
-
-    post = [kw for m, p, kw in client.calls if m == "POST"][0]
-    assert "identity_proof" not in post["json"]
-
-
-def test_guests_revoke_posts_to_revoke_endpoint() -> None:
-    """``guests.revoke(external_id)`` disables the guest's allowlist row via the
-    receiver's revoke endpoint (FR-79)."""
-    from kamiwaza_sdk.services.federations import FederationsAPI
-
-    client = _MockClient()
-    fid = "33333333-3333-3333-3333-333333333333"
-    _stage_guest_federation(client, fid=fid)
-    client.expect(
-        "POST",
-        f"/cluster/federations/{fid}/guests/carol@src/revoke",
-        {"success": True},
-    )
-
-    result = FederationsAPI(client)["ORION"].guests.revoke("carol@src")
-    assert result == {"success": True}
-    assert ("POST", f"/cluster/federations/{fid}/guests/carol@src/revoke") in [
-        (m, p) for m, p, _ in client.calls
-    ]
-
-
-# ---------------------------------------------------------------------------
-# ENG-8213 S8 (design §7.5) — source-side per-target federation credential
-# resolution + attachment on mesh calls to receiver_realm targets
-# ---------------------------------------------------------------------------
-
-
-def test_resolve_federation_credential_from_env() -> None:
-    from kamiwaza_sdk.services.federation_credentials import (
-        resolve_federation_credential,
-    )
-
-    env = {"KAMIWAZA_FEDERATION_CREDENTIAL_ORION": "offline-cred-xyz"}
-    assert resolve_federation_credential("ORION", env=env) == "offline-cred-xyz"
-
-
-def test_resolve_federation_credential_sanitizes_target_name() -> None:
-    from kamiwaza_sdk.services.federation_credentials import (
-        resolve_federation_credential,
-    )
-
-    # non-alnum chars in the federation name map to '_' in the env key.
-    env = {"KAMIWAZA_FEDERATION_CREDENTIAL_ORION_PROD": "cred"}
-    assert resolve_federation_credential("orion-prod", env=env) == "cred"
-
-
-def test_resolve_federation_credential_none_when_unset() -> None:
-    from kamiwaza_sdk.services.federation_credentials import (
-        resolve_federation_credential,
-    )
-
-    assert resolve_federation_credential("ORION", env={}) is None
-
-
-def test_resolve_federation_credential_file_fallback(tmp_path) -> None:
-    import json
-
-    from kamiwaza_sdk.services.federation_credentials import (
-        resolve_federation_credential,
-    )
-
-    cred_file = tmp_path / "creds.json"
-    cred_file.write_text(json.dumps({"ORION": "from-file"}))
-    env = {"KAMIWAZA_FEDERATION_CREDENTIAL_FILE": str(cred_file)}
-    assert resolve_federation_credential("ORION", env=env) == "from-file"
-
-
-def test_resolve_federation_credential_env_beats_file(tmp_path) -> None:
-    import json
-
-    from kamiwaza_sdk.services.federation_credentials import (
-        resolve_federation_credential,
-    )
-
-    cred_file = tmp_path / "creds.json"
-    cred_file.write_text(json.dumps({"ORION": "from-file"}))
-    env = {
-        "KAMIWAZA_FEDERATION_CREDENTIAL_ORION": "from-env",
-        "KAMIWAZA_FEDERATION_CREDENTIAL_FILE": str(cred_file),
-    }
-    assert resolve_federation_credential("ORION", env=env) == "from-env"
-
-
-def test_probe_attaches_federation_credential_header_when_resolved(monkeypatch) -> None:
-    """A mesh call to a receiver_realm target carries the resolved credential in
-    the X-KZ-Federation-Credential header so the receiver validates it against
-    its own federation-<id> realm (design §7.5)."""
-    from kamiwaza_sdk.services.federations import FederationsAPI
-
-    monkeypatch.setenv("KAMIWAZA_FEDERATION_CREDENTIAL_ORION", "offline-cred-xyz")
-    client = _MockClient()
-    client.expect(
-        "GET",
-        "/mesh/ORION/api/cluster/cluster_capabilities",
-        {"system_type": "linux", "os": "linux"},
-    )
-    FederationsAPI(client)["ORION"].probe()
-
-    call = [kw for m, p, kw in client.calls if p.endswith("/cluster_capabilities")][0]
-    assert call["headers"]["X-KZ-Federation-Credential"] == "offline-cred-xyz"
-
-
-def test_probe_omits_federation_credential_header_when_absent(monkeypatch) -> None:
-    from kamiwaza_sdk.services.federations import FederationsAPI
-
-    monkeypatch.delenv("KAMIWAZA_FEDERATION_CREDENTIAL_ORION", raising=False)
-    client = _MockClient()
-    client.expect(
-        "GET",
-        "/mesh/ORION/api/cluster/cluster_capabilities",
-        {"system_type": "linux", "os": "linux"},
-    )
-    FederationsAPI(client)["ORION"].probe()
-
-    call = [kw for m, p, kw in client.calls if p.endswith("/cluster_capabilities")][0]
-    headers = call.get("headers") or {}
-    assert "X-KZ-Federation-Credential" not in headers
